+++ src/modules/config-editor/config-editor.service.ts
@@ -36,7 +36,11 @@ export class ConfigEditorService {
       config.bridge.name = 'Homebridge';
     }
 
-    if (!config.bridge.port) {
+    if (typeof config.bridge.port === 'string') {
+      config.bridge.port = parseInt(config.bridge.port, 10);
+    }
+
+    if (!config.bridge.port || typeof config.bridge.port !== 'number' || config.bridge.port > 65533 || config.bridge.port < 1025) {
       config.bridge.port = Math.floor(Math.random() * (52000 - 51000 + 1) + 51000);
     }